+++ internal/owl/query.go
@@ -10,10 +10,68 @@ import (
 
 	"github.com/graphql-go/graphql/language/ast"
 	"github.com/graphql-go/graphql/language/printer"
+	"go.uber.org/zap"
 )
 
-type Query struct {
-	doc *ast.Document
+func (s *Store) snapshotQuery(query, vars io.StringWriter) error {
+	varDefs := []*ast.VariableDefinition{
+		ast.NewVariableDefinition(&ast.VariableDefinition{
+			Variable: ast.NewVariable(&ast.Variable{
+				Name: ast.NewName(&ast.Name{
+					Value: "insecure",
+				}),
+			}),
+			Type: ast.NewNamed(&ast.Named{
+				Name: ast.NewName(&ast.Name{
+					Value: "Boolean",
+				}),
+			}),
+			DefaultValue: ast.NewBooleanValue(&ast.BooleanValue{
+				Value: false,
+			}),
+		}),
+	}
+
+	loaded, updated, deleted := 0, 0, 0
+	for _, opSet := range s.opSets {
+		if len(opSet.specs) == 0 && len(opSet.values) == 0 {
+			continue
+		}
+		switch opSet.operation.kind {
+		case LoadSetOperation:
+			loaded++
+		case UpdateSetOperation:
+			updated++
+		case DeleteSetOperation:
+			deleted++
+		}
+
+	}
+	s.logger.Debug("snapshot opSets breakdown", zap.Int("loaded", loaded), zap.Int("updated", updated), zap.Int("deleted", deleted))
+
+	q, err := NewQuery("Snapshot", varDefs,
+		[]QueryNodeReducer{
+			reconcileAsymmetry(s),
+			reduceSetOperations(s, vars),
+			reduceSepcs(s),
+			reduceSnapshot(),
+		},
+	)
+	if err != nil {
+		return err
+	}
+
+	text, err := q.Print()
+	if err != nil {
+		return err
+	}
+
+	_, err = query.WriteString(text)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 func reduceSetOperations(store *Store, vars io.StringWriter) QueryNodeReducer {
@@ -445,6 +503,10 @@ func reduceSepcs(store *Store) QueryNodeReducer {
 
 		return doneSelSet, nil
 	}
+}
+
+type Query struct {
+	doc *ast.Document
 }
 
 func NewQuery(name string, varDefs []*ast.VariableDefinition, reducers []QueryNodeReducer) (*Query, error) {
